Predicted trenches
