+++ Content.Shared/_RY/Trenches/TrenchedComponent.cs
@@ -1,9 +1,9 @@
-namespace Content.Server._RY.Trenches;
+namespace Content.Shared._RY.Trenches;
 
 /// <summary>
 /// Added onto an entity who is inside of a trench
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, AutoGenerateComponentState]
 public sealed partial class TrenchedComponent : Component
 {
     [DataField] public bool IsTrenched;
